refactor: move protected email integration test to a separate file
